split up analytics utils into module
