+++ corehq/apps/analytics/management/commands/blocked_hubspot_users_summary.py
@@ -4,7 +4,7 @@ import requests
 from django.conf import settings
 from django.core.management import BaseCommand
 
-from corehq.apps.analytics.utils import (
+from corehq.apps.analytics.utils.hubspot import (
     get_blocked_hubspot_domains,
     MAX_API_RETRIES,
     get_first_conversion_status_for_emails,
