+++ corehq/apps/analytics/management/commands/manually_cleanup_blocked_hubspot_contacts.py
@@ -1,6 +1,6 @@
 from django.core.management import BaseCommand
 
-from corehq.apps.analytics.utils import (
+from corehq.apps.analytics.utils.hubspot import (
     remove_blocked_domain_contacts_from_hubspot,
     remove_blocked_domain_invited_users_from_hubspot,
 )
