+++ corehq/apps/analytics/utils/hubspot.py
@@ -29,19 +29,6 @@ ALLOWED_CONVERSIONS = [
 ]
 
 
-def get_meta(request):
-    return {
-        'HTTP_X_FORWARDED_FOR': request.META.get('HTTP_X_FORWARDED_FOR'),
-        'REMOTE_ADDR': request.META.get('REMOTE_ADDR'),
-    }
-
-
-def analytics_enabled_for_email(email_address):
-    from corehq.apps.users.models import CouchUser
-    user = CouchUser.get_by_username(email_address)
-    return user.analytics_enabled if user else True
-
-
 def is_domain_blocked_from_hubspot(domain):
     return Subscription.visible_objects.filter(
         is_active=True,
@@ -121,10 +108,7 @@ def get_blocked_hubspot_accounts():
     ]
 
 
-def get_instance_string():
-    instance = settings.ANALYTICS_CONFIG.get('HQ_INSTANCE', '')
-    env = '' if instance == 'www' else instance + '_'
-    return env
+
 
 
 def _delete_hubspot_contact(vid, retry_num=0):
